Refactor : AddCardButton 폴더 경로상 위치 변경, 공통 카드 스타일 추출
