+++ src/components/Card/CardItem/CardItem.js
@@ -1,5 +1,7 @@
 import styled from "@emotion/styled";
 
+import Item from "../../style/card";
+
 import isPublic from "./utils/isPublic";
 
 const Card = ({ cardItem }) => {
@@ -20,19 +22,6 @@ const Card = ({ cardItem }) => {
     </Item>
   );
 };
-
-export const Item = styled.div`
-  width: 210px;
-  height: 130px;
-  padding: 15px;
-  border-radius: 5px;
-  background: #d2d2d2;
-  box-shadow: 3px 3px 5px rgba(0, 0, 0, 0.25);
-
-  & + & {
-    margin-top: 20px;
-  }
-`;
 
 const Top = styled.p`
   font-size: 10px;
